fix BQ dependencies and missing client.
